feat: Add support for PostgreSQL database
