+++ pkl/sonr.conf/Hway.pkl
@@ -7,12 +7,11 @@ import "package://pkg.pkl-lang.org/pkl-go/pkl.golang@0.5.0#/go.pkl"
 
 servePort: Int = 3000
 
-configDir: String = "~/.config/hway"
 sqliteFile: String = "hway.db"
 chainId: String = "sonr-testnet-1"
 ipfsGatewayUrl: String = "http://localhost:8080"
 sonrApiUrl: String = "http://localhost:1317"
 sonrGrpcUrl: String = "http://localhost:9090"
 sonrRpcUrl: String = "http://localhost:26657"
-
+psqlDSN: String
